Refactor CardView.tsx
