+++ src/components/View/CardView.tsx
@@ -5,31 +5,14 @@ import { LoadingSpinner } from '../LoadingSpinner';
 import { parseCardMetaView } from '../../utils/Parser';
 import { useMutation } from '@tanstack/react-query';
 import { useParams } from 'react-router-dom';
-import React, { Key, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { ViewHeader } from './ViewHeader';
-import {
-    getUserLanguage,
-    resolveCardBindings,
-    sanitizeBinding,
-} from '../../utils/utils';
+import { getUserLanguage, resolveCardBindings } from '../../utils/utils';
 import { getEntityPropertiesSchema } from '../../temp/SchemaUtils';
-import { List } from './List';
-import { TranslationList } from './TranslationList';
-import { SectionHeading } from './SectionHeading';
+import { CardViewBuilder } from './CardView/CardViewBuilder';
 
 export type DataProps = {
     view: Element;
-};
-
-type ElementAttributesType = {
-    __id: string;
-    [index: string]: string;
-};
-
-type CardElementType = {
-    name: 'Group' | 'List' | 'Element';
-    attributes: ElementAttributesType;
-    [index: string]: unknown;
 };
 
 export const CardView = ({ view }: DataProps) => {
@@ -71,42 +54,6 @@ export const CardView = ({ view }: DataProps) => {
         resolvedSubHeader = SubHeader;
     }
 
-    // Data structure for possible exceptions like Translations table
-    const exceptionElements = new Set<string>();
-    const ExceptionEntityTypes = new Set([
-        'Translation',
-        'TranslationCollection',
-    ]);
-
-    const entityPropertiesAndTypes = new Map<string, string>();
-
-    // Find possible known exceptions from Schema
-    if (EntityPropertySchema) {
-        Object.entries(EntityPropertySchema as DynamicObject).forEach(
-            (property) => {
-                const key = property[0];
-                const val = property[1];
-
-                if (
-                    val.Type === 'List' &&
-                    val.SubType &&
-                    ExceptionEntityTypes.has(val.SubType.Type)
-                ) {
-                    exceptionElements.add(key);
-                }
-
-                entityPropertiesAndTypes.set(key, val.Type);
-            },
-        );
-    }
-
-    const getPath = (bindingExpression: string) => {
-        const path = bindingExpression
-            .replace('{Binding ', '')
-            .replace('}', '');
-        return path.split('.');
-    };
-
     const viewModelSearchOptions = {
         ViewName: viewName,
         ArgumentType: argumentType,
@@ -129,177 +76,6 @@ export const CardView = ({ view }: DataProps) => {
         mutation.mutate(viewModelSearchOptions);
     }, []);
 
-    // console.log(cardData);
-
-    const PrintList = ({ element }: { element: DynamicObject }) => {
-        if (cardData === null) return null;
-        return <List xmlElementTree={element} listData={cardData} />;
-    };
-
-    const PrintElement = ({ element }: { element: DynamicObject }) => {
-        console.log(element);
-        const cardDetails = resolveCardBindings(
-            cardData,
-            element.attributes.Value,
-        );
-
-        const sanitizedBinding = sanitizeBinding(element.attributes.Value);
-        const woEntity = sanitizedBinding.replace('Entity.', '');
-        if (entityPropertiesAndTypes.get(woEntity) === 'Boolean') {
-            return (
-                <div className={`flex flex-col lg:flex-row lg:gap-32`}>
-                    <label
-                        htmlFor={element.attributes.Identifier}
-                        className={`text-sm font-semibold text-ad-grey-800 flex items-center lg:w-1/4`}
-                    >
-                        {element.attributes.Caption}
-                    </label>
-                    <input
-                        id={element.attributes.Identifier}
-                        type={`checkbox`}
-                        checked={cardDetails?.toString() === 'true'}
-                        onChange={() => {
-                            console.log(
-                                `checkbox clicked: ${element.attributes.Caption}`,
-                            );
-                        }}
-                    />
-                </div>
-            );
-        }
-
-        if (cardDetails && entityPropertiesAndTypes.get(woEntity) === 'Date') {
-            return (
-                <div className={`flex flex-col lg:flex-row lg:gap-32`}>
-                    <label
-                        htmlFor={element.attributes.Identifier}
-                        className={`text-sm font-semibold text-ad-grey-800 flex items-center lg:w-1/4`}
-                    >
-                        {element.attributes.Caption}
-                    </label>
-                    <input
-                        id={element.attributes.Identifier}
-                        type={`text`}
-                        defaultValue={new Date(
-                            cardDetails.toString(),
-                        ).toLocaleString('fi-FI', {
-                            year: 'numeric',
-                            month: 'numeric',
-                            day: 'numeric',
-                        })}
-                        className={`flex-1 max-h-12 border border-ad-grey-300 rounded-sm px-2 py-1 hover:border-ad-primary focus:border-ad-primary active:border-ad-primary focus:outline-none`}
-                    />
-                </div>
-            );
-        }
-
-        const isCardDetailsNull =
-            cardDetails === null || cardDetails === undefined;
-
-        const isElementException = getPath(element.attributes.Value).find(
-            (subPath) => exceptionElements.has(subPath),
-        );
-
-        if (isElementException) {
-            if (!Array.isArray(cardDetails) || cardDetails.length === 0)
-                return null;
-
-            return (
-                <div className={`basis-full my-8 col-span-2`}>
-                    <h2 className={`text-lg mb-2 uppercase text-ad-grey-700`}>
-                        {element.attributes.Caption}
-                    </h2>
-                    <TranslationList translations={cardDetails} />
-                </div>
-            );
-        }
-
-        return (
-            <>
-                {!isCardDetailsNull && !Array.isArray(cardDetails) && (
-                    <div className={`flex flex-col lg:flex-row lg:gap-32`}>
-                        <label
-                            htmlFor={element.attributes.Identifier}
-                            className={`text-sm font-semibold text-ad-grey-800 flex items-center lg:w-1/4`}
-                        >
-                            {element.attributes.Caption}
-                        </label>
-                        <input
-                            id={element.attributes.Identifier}
-                            type={typeof cardDetails}
-                            defaultValue={cardDetails.toString()}
-                            className={`flex-1 max-h-12 border border-ad-grey-300 rounded-sm px-2 py-1 hover:border-ad-primary focus:border-ad-primary active:border-ad-primary focus:outline-none`}
-                        />
-                    </div>
-                )}
-            </>
-        );
-    };
-
-    const MapElements = ({
-        elements,
-    }: {
-        elements: Array<CardElementType>;
-    }) => {
-        return (
-            <>
-                {elements.map((element: CardElementType) => {
-                    if (element.name === 'Group')
-                        return (
-                            <PrintGroup
-                                group={element}
-                                key={element.attributes['__id'] as Key}
-                            />
-                        );
-                    else if (element.name === 'List') {
-                        return (
-                            <PrintList
-                                key={element.attributes['__id'] as Key}
-                                element={element}
-                            />
-                        );
-                    } else if (element.name === 'Element')
-                        return (
-                            <PrintElement
-                                key={element.attributes['__id'] as Key}
-                                element={element}
-                            />
-                        );
-                    else {
-                        return (
-                            <p key={element.attributes['__id'] as Key}>
-                                {`Element type ${element.name} not
-                                                yet implemented`}
-                            </p>
-                        );
-                    }
-                })}
-            </>
-        );
-    };
-
-    const PrintGroup = ({ group }: { group: DynamicObject }) => {
-        const [isContentHidden, setIsContentHidden] = useState(false);
-
-        return (
-            group.children?.length && (
-                <>
-                    {group.attributes.Caption && (
-                        <SectionHeading
-                            onClick={() => setIsContentHidden(!isContentHidden)}
-                        >
-                            {group.attributes.Caption}
-                        </SectionHeading>
-                    )}
-
-                    {!isContentHidden && (
-                        <MapElements elements={group.children} />
-                    )}
-                </>
-            )
-        );
-    };
-
     const constructCardView = (parsedCardMetaView: DynamicObject) => {
         return (
             <>
@@ -311,18 +87,11 @@ export const CardView = ({ view }: DataProps) => {
                 )}
                 {parsedCardMetaView && (
                     <div className="px-8 grid lg:grid-cols-2 gap-y-2 gap-x-64 pb-16 lg:max-w-[90%]">
-                        {parsedCardMetaView.children.map(
-                            (XmlNode: DynamicObject) => {
-                                if (XmlNode.name === 'Group') {
-                                    return (
-                                        <PrintGroup
-                                            group={XmlNode}
-                                            key={XmlNode.attributes.__id}
-                                        />
-                                    );
-                                }
-                            },
-                        )}
+                        <CardViewBuilder
+                            elements={parsedCardMetaView.children}
+                            cardData={cardData}
+                            entityPropertySchema={EntityPropertySchema}
+                        />
                     </div>
                 )}
             </>
